+++ src/widgets/index.js
@@ -1,12 +1,5 @@
 import React from 'react';
 
-import Weather from "./Weather/Weather";
-import Clock from "./Clock/Clock";
-
-export const WEATHER    = { type: 'weather', el: <Weather /> };
-export const CLOCK      = { type: 'clock', el: <Clock /> };
-
-
 /** Adding New Widget Form:
  * 1. Import your Widget.
  * 2. Copy the below export and add it to the bottom of the exports.
@@ -14,3 +7,11 @@ export const CLOCK      = { type: 'clock', el: <Clock /> };
  *    Widget.
  */
 // export const WIDGET  = { type: 'KEY', el: '<Widget />' };
+
+import Weather from "./Weather/Weather";
+import Clock from "./Clock/Clock";
+import Calendar from './Calendar/Calendar';
+
+export const WEATHER    = { type: 'weather', el: <Weather /> };
+export const CLOCK      = { type: 'clock', el: <Clock /> };
+export const CALENDAR      = { type: 'calendar', el: <Calendar /> };