+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## March 2, 2021
+
+- Release v0.2.
+- Added keybind (Esc) to close Settings panel.
+- Widgets Added: Calendar
+
 ## February 25, 2021
 
-- Released Version 0.0.1.
+- Released v0.1.
+- Basic features added.
+- Widgets Added: Weather, Clock